+++ src/lib/components/features/transaction/transaction-table-body.svelte
@@ -1,11 +1,15 @@
 <script lang="ts" generics="Item extends { id: string }">
 	import type { Snippet } from 'svelte';
 
+	import { cn } from 'tailwind-variants';
+
 	let {
+		class: className,
 		createrow,
 		data,
 		row
 	}: {
+		class?: string;
 		createrow?: Snippet;
 		data: Item[];
 		row: Snippet<
@@ -16,7 +20,7 @@
 	let editRowId: null | string = $state(null);
 </script>
 
-<div role="rowgroup" class="grid space-y-1.5">
+<div role="rowgroup" class={cn('grid space-y-1.5', className)}>
 	{@render createrow?.()}
 
 	{#each data as item (item.id)}